book: 스프링 부트와 AWS로 혼자 구현하는 웹 서비스 PostsRepositoryTest 에 Controller 관련 기능 테스트가 있어 파일로 분리하였다.
