+++ book/스프링_부트와_AWS로_혼자_구현하는_웹_서비스/project/src/test/java/com/apeltop/domain/posts/PostsRepositoryTest.java
@@ -1,15 +1,10 @@
 package com.apeltop.domain.posts;
 
-import com.apeltop.web.dto.PostsSaveRequestDto;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.List;
@@ -20,41 +15,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class PostsRepositoryTest {
-    @LocalServerPort
-    private int port;
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
     @Autowired
     PostsRepository postsRepository;
 
     @After
     public void cleanup() {
         postsRepository.deleteAll();
-    }
-
-    @Test
-    public void Posts_등록된다() throws Exception {
-        String title = "title";
-        String content = "content";
-
-        PostsSaveRequestDto requestDto = PostsSaveRequestDto
-                .builder()
-                .title(title)
-                .content(content)
-                .build();
-
-        String url = "http://localhost:" + port + "/api/v1/posts";
-
-        ResponseEntity<Long> responseEntity = restTemplate.postForEntity(url, requestDto, Long.class);
-
-        assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(responseEntity.getBody()).isGreaterThan(0L);
-
-        List<Posts> all = postsRepository.findAll();
-        assertThat(all.get(0).getTitle()).isEqualTo(title);
-        assertThat(all.get(0).getContent()).isEqualTo(content);
     }
 
     @Test
